requirements.txt
langchain>=0.1.0
openai>=1.12.0
langchain-google-genai>=0.0.5
langchain-groq>=0.0.1
langchain-community>=0.0.10
yfinance>=0.2.36
ta>=0.11.0
statsmodels>=0.14.1
tensorflow>=2.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.4.0
pmdarima>=2.0.4
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
TA-Lib>=0.4.28